change weight_average type to float
Changed the weight_average type to float and re-write the last task procedure

+++ 0x00-MySQL_Advanced/100-average_weighted_score.sql
@@ -9,7 +9,7 @@ DELIMITER $
 CREATE PROCEDURE ComputeAverageWeightedScoreForUser(IN usr_id INT)
 -- begin procedure calculations
 BEGIN
-    DECLARE weight_average INT;
+    DECLARE weight_average FLOAT;
     DECLARE total_weight_for_user INT;
     
     SELECT SUM(weight) FROM corrections
